client: C++ compatibility for cl_avi

+++ code/client/cl_avi.c
@@ -556,7 +556,7 @@ qboolean CL_OpenAVIForWriting( const char *fileName, const char *pipeFormat, qbo
 	afd.a.bits = dma.samplebits;
 	afd.a.sampleSize = (afd.a.bits * afd.a.channels) / 8;
 
-	afd.audioFrameSize = ceil( (float)(afd.a.rate * afd.a.sampleSize) / (float)afd.frameRate );
+	afd.audioFrameSize = ceil( (double)((float)(afd.a.rate * afd.a.sampleSize) / (float)afd.frameRate) );
 
 	if ( Cvar_VariableIntegerValue( "s_initsound" ) == 0 )
 	{
